+++ src/checks/application/mod.rs
@@ -24,4 +24,5 @@ pub mod s7comm;
 pub mod smtp;
 pub mod snmp;
 pub mod srvloc;
+pub mod ssh;
 pub mod tls;