+++ src/errors/application/mod.rs
@@ -27,6 +27,7 @@ pub mod s7comm;
 pub mod smtp;
 pub mod snmp;
 pub mod srvloc;
+pub mod ssh;
 pub mod tls;
 
 /// Errors related to parsing an `Application`